Split ServerSettingsCard into Card and CardItem to re-use in GetStartedWithThundermail
